% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm_ep_em}
\alias{lmm_ep_em}
\title{Parallel LMM}
\usage{
lmm_ep_em(y, X, Z, subject, beta, R, D, sigma, nrm = 100, maxiter = 500,
  cores = 8, verbose = TRUE)
}
\arguments{
\item{y}{matrix of responses with observations/subjects on column and repeats for each observation/subject on rows. It is (m x n) dimensional.}

\item{X}{observed design matrices for fixed effects. It is (m*n x p) dimensional.}

\item{Z}{observed design matrices for random effects. It is (m*n x q) dimensional.}

\item{beta}{fixed effect estimation vector with length p.}

\item{R}{variance-covariance matrix of residuals.}

\item{D}{variance-covariance matrix of random effects.}

\item{cores}{the number of cores. Why not to use maximum?!}

\item{verbose}{boolean that defaults to print iteration context}
}
\description{
description
}
