% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm.ep.em}
\alias{lmm.ep.em}
\title{Title}
\usage{
lmm.ep.em(y, X, Z, subject, beta, R, D, sigma, nrm = 100, maxiter = 500,
  cores = 8)
}
\arguments{
\item{y}{matrix of responses with observations/subjects on column and repeats for each observation/subject on rows. It is (m $\times$ n) dimensional.}

\item{X}{desc1}

\item{Z}{desc1}

\item{beta}{desc1}

\item{R}{desc2}

\item{D}{desc3}

\item{cores}{desc4}
}
\description{
description
}
