% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm.ep.em}
\alias{lmm.ep.em}
\title{Title}
\usage{
lmm.ep.em(y, X, Z, beta, R, D, sigma, cores = 3, nu = 3, nrm = 100,
  maxiter = 500, first_parallel = FALSE, second_parallel = FALSE)
}
\arguments{
\item{y}{desc1}

\item{X}{desc1}

\item{Z}{desc1}

\item{beta}{desc1}

\item{R}{desc2}

\item{D}{desc3}

\item{cores}{desc4}
}
\description{
description
}
